+++ src/Credfeto.Extensions.Linq.Tests/EnumerableExtensionsTests.cs
@@ -7,7 +7,7 @@ namespace Credfeto.Extensions.Linq.Tests;
 public sealed class EnumerableExtensionsTests : TestBase
 {
     [Fact]
-    public void ShouldRemoveNulls()
+    public void ShouldRemoveNullsClass()
     {
         string?[] input =
         {
@@ -26,6 +26,30 @@ public sealed class EnumerableExtensionsTests : TestBase
                          "A",
                          "B",
                          "C"
+                     },
+                     actual: output);
+    }
+
+    [Fact]
+    public void ShouldRemoveNullsValue()
+    {
+        int?[] input =
+        {
+            1,
+            null,
+            2,
+            null,
+            3
+        };
+
+        int[] output = input.RemoveNulls()
+                            .ToArray();
+
+        Assert.Equal(new[]
+                     {
+                         1,
+                         2,
+                         3
                      },
                      actual: output);
     }