+++ src/Credfeto.Extensions.Linq/EnumerableExtensions.cs
@@ -15,4 +15,16 @@ public static class EnumerableExtensions
             }
         }
     }
+
+    public static IEnumerable<TItemType> RemoveNulls<TItemType>(this IEnumerable<TItemType?> source)
+        where TItemType : struct
+    {
+        foreach (TItemType? item in source)
+        {
+            if (item.HasValue)
+            {
+                yield return item.Value;
+            }
+        }
+    }
 }